Rename data type file and add new type
