+++ src/components/DropdownMenu/type.ts
@@ -1,6 +1,6 @@
 import Position from "../../types/position.type";
 import ImageBorder from "../../types/imageBorder.type";
-import VersionData from "../../types/versionData.type";
+import { VersionData } from "../../types/pokemonData.type";
 import { MouseEvent } from "react";
 
 type DropdownMenuProps = {
