+++ src/components/GameVersionCard/type.ts
@@ -1,6 +1,6 @@
 import Difficulty from "../../types/difficulty.type";
 import GameVersion from "../../types/gameVersion.type";
-import VersionData from "../../types/versionData.type";
+import { VersionData } from "../../types/pokemonData.type";
 
 type GameVersionCardProps = {
   difficulty: Difficulty;
