+++ src/pages/Game/type.ts
@@ -1,7 +1,7 @@
 import FormatTime from "../../types/formatTime.type";
 import GameVersion from "../../types/gameVersion.type";
 import ValidationData from "../../types/validationData.type";
-import VersionData from "../../types/versionData.type";
+import { VersionData } from "../../types/pokemonData.type";
 
 type GameProps = {
   gameData: VersionData;
