--- src/types/versionData.type.ts
@@ -1,6 +1,0 @@
-type VersionData = {
-  pokemonNames: Array<string>;
-  imageDimensions: { width: number; height: number };
-};
-
-export default VersionData;
